crudにboardの関数を移動
